requirements.txt
aistudio-sdk==0.3.8
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.46
blinker==1.9.0
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
easyocr==1.7.2
filelock==3.20.0
Flask==3.1.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
modelscope==1.31.0
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.0
paddlepaddle==3.2.0
paddlex==3.3.5
pandas==2.3.3
pillow==12.0.0
prettytable==3.16.0
protobuf==6.30.2
psutil==7.1.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tifffile==2025.10.16
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
